requirements.txt
streamlit
requests
beautifulsoup4
google-cloud-spanner
python-dotenv
openai
